[unit_prop] fix issue with conflicting unit_prop
Summary: If I have
A && B
A && !B

When I unit prop with the assignment A=False, first I infer that B=True. Then when I look at the second clause, I have a conflict.

Unfortunately, we had deferred adding inferred values to the assignment, meaning the second clause was evaluated with just A=False, not [A=False, B=True]. This meant we didn't get a conflict.

Test Plan: the dimacs examples no longer panic