[store] move to ClauseRef abstraction
Summary: The existing system of Clauses references is fine, as long as we never hold the references between iterations of unit prop or evaluation. This is broadly possible, if we're careful. But it's a bit inconvenient.

An alternate approach is to store a "view" - in this case, an offset into the ClausesList array. This allows us to mutate the underlying ClausesList (by, for example, adding new clauses) without dealing with all the lifetime nonsence. OTOH, we do need some extra accounting when it comes to looking up the literals in a clause, in case they've been removed or whatnot.

So, overall changes:
- Introduce ClauseRef, a reference to a Clause in a ClauseList. For short (unit, pair) clauses, we inline the literals.
- Introduce various utilities to build, look up and iterate over ClauseRefs.
- Introduce ClauseStore, to hold the ClauseList and ClauseIndex. It is responsible for keeping the two stores in sync
- Introduce ClauseIndexView, which binds the ClauseStore and ClauseIndex into a lightweight read only view of the index
- Migrate everything to use the new ClauseRef abstraction

Test Plan: Added a couple of unit tests